Runner: store the updated formatted output of clang-format when necessary.
This commit adds a better test case which includes a file that should be reformatted, and a file that should not. It also contains a .clang-format which standardizes the output in order to make the test more reliable (though it definitely needs a more flexible interface for selecting which version of clang-format should be used).